pyproject.toml
[tool.poetry]
name = "whisper-music"
version = "0.1.0"
description = ""
authors = ["aidanjrauscher <ajrscomputer@gmail.com>"]
readme = "README.md"
packages = [{include = "whisper_music"}]

[tool.poetry.dependencies]
python = "^3.11"
openai = "^0.27.6"
langchain = "^0.0.158"
python-dotenv = "^1.0.0"
speechrecognition = "^3.10.0"
pyaudio = "^0.2.13"
pyautogui = "^0.9.53"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
